feat: allow reinstalling apps, new app id format
fixes possible problems if accidently removing an entry

allows transitioning to new IDs since GioApplication rejects app ids
that contain segments which starts with a number so an app id like
io.github.zaedus.6abc...db would be rejected. This was also a problem
with CSS classes, but sadly is not so easily fixable.